Move `AppointmentOutcomeForm` object
I find it easier to work with types alongside usecases.
