+++ integration_tests/mockApis/appointmentForms.ts
@@ -1,6 +1,8 @@
 import { SuperAgentRequest } from 'superagent'
-import { AppointmentOutcomeForm } from '../../server/@types/user-defined'
-import { APPOINTMENT_UPDATE_FORM_TYPE } from '../../server/services/forms/appointmentFormService'
+import {
+  APPOINTMENT_UPDATE_FORM_TYPE,
+  AppointmentOutcomeForm,
+} from '../../server/services/forms/appointmentFormService'
 import { stubFor } from './wiremock'
 
 export default {
