+++ server/pages/appointments/chooseSupervisorPage.test.ts
@@ -6,7 +6,7 @@ import supervisorSummaryFactory from '../../testutils/factories/supervisorSummar
 import CheckAppointmentDetailsPage from './checkAppointmentDetailsPage'
 import * as Utils from '../../utils/utils'
 import appointmentOutcomeFormFactory from '../../testutils/factories/appointmentOutcomeFormFactory'
-import { AppointmentOutcomeForm } from '../../@types/user-defined'
+import { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import ChooseSupervisorPage, { AppointmentDetailsQuery } from './chooseSupervisorPage'
 import providerTeamSummaryFactory from '../../testutils/factories/providerTeamSummaryFactory'
 import Offender from '../../models/offender'
