+++ server/pages/appointments/chooseSupervisorPage.ts
@@ -1,10 +1,6 @@
 import { ProviderTeamSummaryDto, SupervisorSummaryDto } from '../../@types/shared'
-import {
-  AppointmentOutcomeForm,
-  AppointmentUpdateQuery,
-  GovUkSelectOption,
-  ValidationErrors,
-} from '../../@types/user-defined'
+import { AppointmentUpdateQuery, GovUkSelectOption, ValidationErrors } from '../../@types/user-defined'
+import { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import GovUkSelectInput from '../../forms/GovUkSelectInput'
 import BaseAppointmentUpdatePage from './baseAppointmentUpdatePage'
 import { AppointmentFormPage } from './pathMap'
