+++ server/pages/appointments/confirmPage.test.ts
@@ -4,7 +4,8 @@ import appointmentFactory from '../../testutils/factories/appointmentFactory'
 import sessionFactory from '../../testutils/factories/sessionFactory'
 import ConfirmPage from './confirmPage'
 import * as Utils from '../../utils/utils'
-import { AppointmentOutcomeForm, YesOrNo } from '../../@types/user-defined'
+import { YesOrNo } from '../../@types/user-defined'
+import { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import appointmentOutcomeFormFactory from '../../testutils/factories/appointmentOutcomeFormFactory'
 import { contactOutcomeFactory } from '../../testutils/factories/contactOutcomeFactory'
 import DateTimeFormats from '../../utils/dateTimeUtils'
